updated configs for the newest version of android studio
for lombok you gotta go on the website marketplace for intellij, manually download the lombok plugin as a zip, dearchive it and copy paste it in: Android Studio -> file location = will be the bin folder -> go to parent folder -> enter the plugins folder and paste it there. For it to work, all Android Studio must be closed. After paste, open the AS and that would be it